moving logger options to a new folder
